+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## v1.0.0-alpha2 - 2021-01-11
+
 ### Added
 
 - Add support for messages, they can be sent between actors and the environment.